requirements.txt
alembic==1.3.0
Mako==1.1.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
protobuf==3.10.0
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.1.2
requests==2.7.0
six==1.12.0
SQLAlchemy==1.3.10
